Do not redo pro link of already correct

+++ lstchain/scripts/onsite/onsite_create_drs4_time_file.py
@@ -98,9 +98,9 @@ def main():
     if pro_symlink:
         pro = "pro"
         pro_dir = f"{output_dir}/../{pro}"
-        if os.path.exists(pro_dir):
+        if os.path.exists(pro_dir) and os.readlink(pro_dir) is not output_dir:
             os.remove(pro_dir)
-        os.symlink(prod_id, pro_dir)
+            os.symlink(prod_id, pro_dir)
         print("\n--> Use symbolic link pro")
     else:
         pro = prod_id
